Se arregló error al eliminar mascota sin foto

+++ src/main/java/com/z/vetdbz/controllers/PetController.java
@@ -84,11 +84,14 @@ public class PetController {
     public String petDelete(@PathVariable Integer id) {
         // Eliminar foto
         Pet pet = iPet.petById(id);
-        Path urlPath = Paths.get("uploads").resolve(pet.getUrlImg()).toAbsolutePath();
-        File photo = urlPath.toFile();
 
-        if (photo.exists() && photo.canRead()) {
-            photo.delete();
+        if (pet.getUrlImg() != null) {
+            Path urlPath = Paths.get("uploads").resolve(pet.getUrlImg()).toAbsolutePath();
+            File photo = urlPath.toFile();
+
+            if (photo.exists() && photo.canRead()) {
+                photo.delete();
+            }
         }
 
         iPet.deletePet(id);
